Use API to search for customers in academy and remove code that was contacting the db
