+++ .env.sample
@@ -49,6 +49,8 @@
 SINGLEVIEW_DB=postgresql://singleview_user:@localhost:10102/hnsingleview 
 /hn-single-view-api/test/SHARED_PLAN_BASE_URL=
 /hn-single-view-api/test/VULNERABILITIES_BASE_URL=
+/hn-single-view-api/test/ACADEMY_API_BASE_URL=http://localhost:8090
+/hn-single-view-api/test/ACADEMY_API_API_KEY=super-secret-api-key
 ################
 
 /hn-single-view-api/SENTRY_DSN=
