+++ package.json
@@ -36,7 +36,7 @@
     "start-jigsaw-docs": "sls offline start -s dev --port 3015 --noAuth --config serverless-jigsaw-docs.yml",
     "debug-test": "export SLS_DEBUG=* && node --inspect ./node_modules/serverless/bin/serverless offline -s test --port 3010 --noAuth --config serverless-api.yml",
     "unit-tests": "jest --testPathIgnorePatterns=api/test/integration --coverage",
-    "start-test": "sls offline start -s test --port 3010 --noAuth --config serverless-api.yml & node ./api/test/FakeJigsawAndW2.js",
+    "start-test": "sls offline start -s test --port 3010 --noAuth --config serverless-api.yml & node ./api/test/FakeJigsawAndW2.js & node ./api/test/FakePlatformApis.js",
     "integration-tests": "jest --testPathPattern=api/test/integration --runInBand",
     "ci": "start-server-and-test start-test http://localhost:3010/customers?firstName=jon integration-tests",
     "lint": "eslint --fix --ext js api/lib && echo 'Lint complete.'"
